refactor: full user resource
