+++ rgdps/adapters/__init__.py
@@ -7,3 +7,4 @@ from .pubsub import RedisPubsubRouter
 from .storage import AbstractStorage
 from .storage import LocalStorage
 from .storage import S3Storage
+from .meilisearch import MeiliSearchClient
