+++ rgdps/common/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from . import cache
+from . import colour
 from . import context
 from . import data_utils
 from . import gd_logic
